add get profile api

+++ pkg/user/jwt/jwt.go
@@ -21,12 +21,12 @@ func CreateToken(id uint, username string) (string, error) {
 	return token.SignedString([]byte(Secret))
 }
 
-func GetUser(c echo.Context) (id int64, username string) {
+func GetUser(c echo.Context) (id uint, username string) {
 	user := c.Get("user").(*jwt.Token)
 	log.Infof("access token: %s", user.Raw)
 	claims := user.Claims.(jwt.MapClaims)
 
-	id = int64(claims["id"].(float64))
+	id = uint(claims["id"].(float64))
 	username = claims["name"].(string)
 	return
 }
